third_party/blink/web_tests/external/wpt/html/dom/reflection-text-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] a.referrerPolicy: setAttribute() to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] a.referrerPolicy: setAttribute() to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] a.referrerPolicy: setAttribute() to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] a.referrerPolicy: setAttribute() to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] a.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] a.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] a.referrerPolicy: IDL set to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] a.referrerPolicy: IDL set to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] a.referrerPolicy: IDL set to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] a.referrerPolicy: IDL set to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] a.referrerPolicy: IDL set to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] a.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
Harness: the test ran to completion.
